requirements.txt
Flask
Flask-Cors
tensorflow
numpy
Pillow
requests
beautifulsoup4
google-api-python-client
python-dotenv